backend/requirements.txt
# 🎬 Auto Video Producer - Backend Dependencies

# Core Flask
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7

# Database
SQLAlchemy==1.4.53

# AI APIs
openai==1.51.0
google-generativeai==0.8.3
anthropic==0.40.0

# HTTP Requests
requests==2.31.0
urllib3==2.2.0

# YouTube APIs
google-api-python-client==2.108.0
isodate==0.6.1
yt-dlp>=2025.9.5

# Audio/Video Processing
moviepy==2.1.2
soundfile==0.13.1
numpy==1.26.4
Pillow==10.4.0
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
tqdm==4.66.1

# Development
pytest==8.3.3
pytest-flask==1.3.0
black==24.8.0
flake8==7.1.1

# TTS Services
elevenlabs==0.2.24

# Additional Services
together==0.2.7
gtts==2.4.0

# Image Processing
requests-toolbelt==1.0.0

# Optional (install if needed)
# Additional AI providers can be added here
